+++ src/domain/player/accumulator.rs
@@ -13,10 +13,10 @@ pub type AccumulatorKey = Uuid;
 pub struct PlayerAccumulator {
     pub id: AccumulatorKey,
     pub player_id: PlayerKey,
-    pub food: i32,
-    pub wood: i32,
-    pub stone: i32,
-    pub gold: i32,
+    pub food: i64,
+    pub wood: i64,
+    pub stone: i64,
+    pub gold: i64,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
